+++ inst/examples/ex-14c-plot.R
@@ -6,7 +6,7 @@ cal <- c14_calibrate(
 )
 
 ## Specify calendar
-plot(cal, calendar = BP(), flip = TRUE)
+plot(cal, calendar = b2k(), flip = TRUE)
 
 ## HDR intervals (default)
 plot(cal, interval = "hdr", level = 0.95)
@@ -15,11 +15,4 @@ plot(cal, interval = "hdr", level = 0.95)
 plot(cal, interval = "credible", level = 0.95)
 
 ## No intervals
-plot(cal, interval = NULL)
-
-## Intervals only
-plot(cal, density = FALSE, level = 0.68, lwd = 5)
-plot(cal, density = FALSE, level = 0.95, lwd = 5)
-
-## Change colors
-plot(cal[, 1, ], col.interval = "red")
+plot(cal, level = 0)